+++ @ornikar/eslint-config-react/rules/react-native.js
@@ -1,0 +1,24 @@
+'use strict';
+
+module.exports = {
+  env: {
+    browser: true,
+  },
+
+  globals: {
+    __DEV__: true,
+  },
+
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.ios.js', '.android.js'],
+      },
+    },
+  },
+
+  rules: {
+    'react/prefer-stateless-function': 'off',
+    'react/no-unescaped-entities': 'off',
+  },
+};