+++ @ornikar/eslint-config-typescript-react/tests/react-native/MyComponent.ios.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import type { ReactElement } from 'react';
+
+export function MyComponent(): ReactElement {
+  return <></>;
+}