+++ @ornikar/eslint-config-typescript/tests/no-unused-vars.ts
@@ -1,4 +1,4 @@
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const foo = someFunction();
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
+// eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/restrict-plus-operands
 const bar = a + 1;