+++ @ornikar/eslint-config-babel/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [14.2.0](https://github.com/ornikar/eslint-configs/compare/v14.1.0...v14.2.0) (2021-06-22)
+
+**Note:** Version bump only for package @ornikar/eslint-config-babel
+
+
+
+
+
+# [14.1.0](https://github.com/ornikar/eslint-configs/compare/v14.0.1...v14.1.0) (2021-06-03)
+
+**Note:** Version bump only for package @ornikar/eslint-config-babel
+
+
+
+
+
+## [14.0.1](https://github.com/ornikar/eslint-configs/compare/v14.0.0...v14.0.1) (2021-05-03)
+
+**Note:** Version bump only for package @ornikar/eslint-config-babel
+
+
+
+
+
 # [14.0.0](https://github.com/ornikar/eslint-configs/compare/v13.3.0...v14.0.0) (2021-04-26)