+++ @ornikar/eslint-config-react/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.2.1](https://github.com/ornikar/eslint-configs/compare/v14.2.0...v14.2.1) (2021-06-29)
+
+
+### Bug Fixes
+
+* **eslint-config-typescript-react:** fix import of react-native config [no issue] ([#57](https://github.com/ornikar/eslint-configs/issues/57)) ([274039c](https://github.com/ornikar/eslint-configs/commit/274039c7e8a184197893d838b36e11aa0908e301))
+
+
+
+
+
+# [14.2.0](https://github.com/ornikar/eslint-configs/compare/v14.1.0...v14.2.0) (2021-06-22)
+
+**Note:** Version bump only for package @ornikar/eslint-config-react
+
+
+
+
+
+# [14.1.0](https://github.com/ornikar/eslint-configs/compare/v14.0.1...v14.1.0) (2021-06-03)
+
+
+### Features
+
+* add rules for eslint meeting ARCH-1376 ([#50](https://github.com/ornikar/eslint-configs/issues/50)) ([15ed376](https://github.com/ornikar/eslint-configs/commit/15ed3761d395562b24cbd3723de8d1055e99a42d))
+
+
+
+
+
 # [14.0.0](https://github.com/ornikar/eslint-configs/compare/v13.3.0...v14.0.0) (2021-04-26)