+++ CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.2.1](https://github.com/ornikar/eslint-configs/compare/v14.2.0...v14.2.1) (2021-06-29)
+
+
+### Bug Fixes
+
+* **eslint-config-typescript-react:** fix import of react-native config [no issue] ([#57](https://github.com/ornikar/eslint-configs/issues/57)) ([274039c](https://github.com/ornikar/eslint-configs/commit/274039c7e8a184197893d838b36e11aa0908e301))
+
+
+
+
+
+# [14.2.0](https://github.com/ornikar/eslint-configs/compare/v14.1.0...v14.2.0) (2021-06-22)
+
+
+### Features
+
+* enable three typescript-eslint rules ([#56](https://github.com/ornikar/eslint-configs/issues/56)) ([3096807](https://github.com/ornikar/eslint-configs/commit/3096807964d6916bce44e57d871d43bd46b51c14))
+
+
+
+
+
+# [14.1.0](https://github.com/ornikar/eslint-configs/compare/v14.0.1...v14.1.0) (2021-06-03)
+
+
+### Bug Fixes
+
+* **eslint-config-typescript:** better typescript eslint config [no issue] ([#39](https://github.com/ornikar/eslint-configs/issues/39)) ([5602a6b](https://github.com/ornikar/eslint-configs/commit/5602a6b1b595daa7c5d84a30c6fbc8bc6d6321b8))
+
+
+### Features
+
+* add rules for eslint meeting ARCH-1376 ([#50](https://github.com/ornikar/eslint-configs/issues/50)) ([15ed376](https://github.com/ornikar/eslint-configs/commit/15ed3761d395562b24cbd3723de8d1055e99a42d))
+
+
+
+
+
+## [14.0.1](https://github.com/ornikar/eslint-configs/compare/v14.0.0...v14.0.1) (2021-05-03)
+
+
+### Bug Fixes
+
+* **eslint-config:** node restricted-syntax [no issue] ([#37](https://github.com/ornikar/eslint-configs/issues/37)) ([94ec626](https://github.com/ornikar/eslint-configs/commit/94ec6264582bd49e6a9346e87af52a23bc0fd4df))
+
+
+
+
+
 # [14.0.0](https://github.com/ornikar/eslint-configs/compare/v13.3.0...v14.0.0) (2021-04-26)